content/media/webspeech/synth/test/test_speech_simple.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=650295
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 650295: Web Speech API check all classes are present</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body onunload="synthCleanup();">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=650295">Mozilla Bug 650295</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 525444 **/

SimpleTest.waitForExplicitFinish();

synthAddVoice('TestSpeechServiceWithAudio', 'Male 1', 'en-GB', true);

var gotStartEvent = false;
var gotBoundaryEvent = false;
var utterance = new SpeechSynthesisUtterance("Hello, world!");
utterance.addEventListener('start', function(e) {
  ok(speechSynthesis.speaking, "speechSynthesis is speaking.");
  ok(!speechSynthesis.pending, "speechSynthesis has no other utterances queued.");
  gotStartEvent = true;
});

utterance.addEventListener('end', function(e) {
  ok(!speechSynthesis.speaking, "speechSynthesis is not speaking.");
  ok(!speechSynthesis.pending, "speechSynthesis has no other utterances queued.");
  ok(gotStartEvent, "Got 'start' event.");
  info('end ' + e.elapsedTime);
  synthCleanup();
  SimpleTest.finish();
});

speechSynthesis.speak(utterance);
ok(!speechSynthesis.speaking, "speechSynthesis is not speaking yet.");
ok(speechSynthesis.pending, "speechSynthesis has an utterance queued.");

</script>
</pre>
</body>
</html>
